+++ doc/tutorials/my_first_widget/materials/display_output.py
@@ -4,7 +4,7 @@ from ewokscore.task import Task
 from silx.gui import qt
 from silx.gui.plot import Plot1D
 
-from ewoksorange.bindings.owwidgets import OWEwoksWidgetOneThread
+from ewoksorange.gui.owwidgets.threaded import OWEwoksWidgetOneThread
 
 
 class ClipDataTask(