+++ src/ewoksorange/gui/concurrency/base.py
@@ -2,7 +2,6 @@ import logging
 from typing import Optional
 from typing import Type
 
-from AnyQt.QtCore import QThread
 from ewokscore import TaskWithProgress
 from ewokscore.task import Task
 from ewokscore.task import TaskInputError
@@ -76,29 +75,3 @@ class TaskExecutor:
     @property
     def current_task(self) -> Optional[Task]:
         return self.__task
-
-
-class ThreadedTaskExecutor(QThread, TaskExecutor):
-    """Create and execute an Ewoks task in a dedicated thread."""
-
-    def run(self) -> None:
-        self.execute_task()
-
-    def stop(self, timeout: Optional[float] = None, wait: bool = False) -> None:
-        """Stop the current thread"""
-        self.blockSignals(True)
-        if wait:
-            if timeout:
-                self.wait(timeout * 1000)
-            else:
-                self.wait()
-        if self.isRunning():
-            self.quit()
-
-    def cancel_running_task(self):
-        """
-        cancel current processing.
-        The targetted EwoksTask must have implemented the 'cancel' function
-        """
-        if self.current_task is not None:
-            self.current_task.cancel()