+++ src/orangecontrib/ewokstest/sumtask.py
@@ -1,7 +1,7 @@
-from ewoksorange.bindings import OWEwoksWidgetNoThread
-from ewoksorange.gui.orange_imports import Input
-from ewoksorange.gui.orange_imports import Output
-from ewoksorange.gui.simpletypesmixin import IntegerAdderMixin
+from ewoksorange.gui.owwidgets.nothread import OWEwoksWidgetNoThread
+from ewoksorange.gui.owwidgets.orange_imports import Input
+from ewoksorange.gui.owwidgets.orange_imports import Output
+from ewoksorange.gui.widgets.simple_types_mixin import IntegerAdderMixin
 from ewoksorange.tests.examples.tasks import SumTaskTest
 
 __all__ = ["OWSumTaskTest"]