+++ CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Deprecated
+
+- Module `ewoksorange.bindings.owwidgets`.
+- Module `ewoksorange.bindings.progress`.
+- Module `ewoksorange.bindings.taskexecutor_queue`.
+- Module `ewoksorange.bindings.taskexecutor`.
+- Module `ewoksorange.gui.data_viewer`.
+- Module `ewoksorange.gui.orange_imports`.
+- Module `ewoksorange.gui.parameterform`.
+- Module `ewoksorange.gui.qtsignals`.
+- Module `ewoksorange.gui.simpletypesmixin`.
+
+## Removed
+
+- Module `ewoksorange.oasys_patch`.
+
 ## [2.1.0] - 2025-10-23
 
 ### Added